new tests and fastapi working

+++ python/fastapi/sql/app.py
@@ -1,6 +1,8 @@
 import os
+import traceback
 from typing import Optional
 from fastapi import FastAPI, Request, status
+from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import RequestValidationError
 from fastapi.responses import JSONResponse
 from sqlmodel import create_engine, SQLModel, Session, Field, select
@@ -15,23 +17,56 @@ class Crudie(SQLModel, table=True):
     data: int
 
 
-@app.on_event("startup")
-def on_startup():
-    SQLModel.metadata.create_all(engine)
-
+class InvalidServiceKey(Exception):
+    pass
 
 
 @app.exception_handler(RequestValidationError)
-def validation_exception_handler(request: Request, exc: RequestValidationError):
+def validation_request_handler(
+    request: Request, exc: RequestValidationError
+) -> JSONResponse:
+    traceback.print_exc()
     return JSONResponse(
         status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-        content={"defailt": exc.errors(), "body": exc.body},
+        content=jsonable_encoder({"detail": exc.errors(), "body": exc.body}),
+    )
+
+
+@app.exception_handler(InvalidServiceKey)
+def invalid_service_key_handler(
+    request: Request, exc: InvalidServiceKey
+) -> JSONResponse:
+    traceback.print_exc()
+    loc_type = "query" if request.method in ("GET", "DELETE") else "body"
+    return JSONResponse(
+        status_code=status.HTTP_428_PRECONDITION_REQUIRED,
+        content=jsonable_encoder(
+            {
+                "details": [
+                    {
+                        "type": "invalid",
+                        "loc": [loc_type, "service_key"],
+                        "msg": "service_key was not valid",
+                    }
+                ],
+                "body": str(request.body()),
+            }
+        ),
     )
 
 
 @app.post("/create", response_model=Crudie)
-def create(service_key: str, data: int):
-    crudie = Crudie(service_key=service_key, data=data)
+def create(crudie: Crudie):
+    if not crudie.service_key:
+        raise RequestValidationError(
+            errors=[
+                {
+                    "type": "invalid",
+                    "loc": ["body", "service_key"],
+                    "msg": "missing service_key",
+                }
+            ]
+        )
     with Session(engine) as session:
         session.add(crudie)
         session.commit()
@@ -49,14 +84,23 @@ def read(service_key: str):
 
 
 @app.put("/update", response_model=Crudie)
-def update(service_key: str, data: int):
-    crudie = Crudie(service_key=service_key, data=data)
+def update(crudie: Crudie):
+    if not crudie.service_key:
+        raise RequestValidationError(
+            errors=[
+                {
+                    "type": "invalid",
+                    "loc": ["body", "service_key"],
+                    "msg": "missing service_key",
+                }
+            ]
+        )
     with Session(engine) as session:
         result = session.exec(
             select(Crudie).where(Crudie.service_key == crudie.service_key)
         ).first()
         if not result:
-            raise
+            raise InvalidServiceKey("service_key not in database")
         result.data = crudie.data
         session.add(result)
         session.commit()
@@ -70,6 +114,8 @@ def delete(service_key: str):
         result = session.exec(
             select(Crudie).where(Crudie.service_key == service_key)
         ).first()
+        if not result:
+            raise InvalidServiceKey("service_key not in database")
         session.delete(result)
         session.commit()
         return result
